Added testing module and updated renderer tests for crate_universe
