+++ crate_universe/src/renderer.rs
@@ -377,8 +377,7 @@ impl Renderer {
         Ok(())
     }
 
-    // TODO: Public for `tests/renderer.rs`, unit tests should added here instead
-    pub fn render_workspaces<Out: Write>(&self, output: &mut Out) -> anyhow::Result<()> {
+    fn render_workspaces<Out: Write>(&self, output: &mut Out) -> anyhow::Result<()> {
         let mut context = tera::Context::new();
         context.insert("lockfile_hash", &self.context.hash);
         context.insert("crates", &self.context.transitive_renderable_packages);
@@ -396,8 +395,7 @@ impl Renderer {
         Ok(())
     }
 
-    // TODO: Public for `tests/renderer.rs`, unit tests should added here instead
-    pub fn render_helper_functions<Out: Write>(&self, output: &mut Out) -> anyhow::Result<()> {
+    fn render_helper_functions<Out: Write>(&self, output: &mut Out) -> anyhow::Result<()> {
         let mut crate_repo_names_inner = BTreeMap::new();
         crate_repo_names_inner.extend(&self.context.member_packages_version_mapping.normal);
         crate_repo_names_inner.extend(&self.context.member_packages_version_mapping.build);
@@ -527,4 +525,286 @@ enum Kind {
     Dev,
     Build,
     ProcMacro,
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::testing;
+
+    use indoc::indoc;
+    use maplit::{btreemap, btreeset};
+
+    fn mock_renderer(git: bool) -> Renderer {
+        Renderer::new(
+            RenderConfig {
+                repo_rule_name: String::from("rule_prefix"),
+                crate_registry_template: String::from(
+                    "https://crates.io/api/v1/crates/{crate}/{version}/download",
+                ),
+                rules_rust_workspace_name: String::from("rules_rust"),
+            },
+            String::from("598"),
+            vec![testing::lazy_static_crate_context(git)],
+            Dependencies {
+                normal: BTreeMap::new(),
+                build: BTreeMap::new(),
+                dev: BTreeMap::new(),
+            },
+            BTreeMap::new(),
+        )
+    }
+
+    #[test]
+    fn render_http_archive() {
+        let renderer = mock_renderer(false);
+
+        let mut output = Vec::new();
+
+        renderer
+            .render_workspaces(&mut output)
+            .expect("Error rendering");
+
+        let output = String::from_utf8(output).expect("Non-UTF8 output");
+
+        // TODO: Don't unconditionally load new_git_repository and http_archive
+        let expected_repository_rule = indoc! { r#"
+            # rules_rust crate_universe file format 1
+            # config hash 598
+
+            load("@bazel_tools//tools/build_defs/repo:git.bzl", "new_git_repository")
+            load("@bazel_tools//tools/build_defs/repo:http.bzl", "http_archive")
+
+            def pinned_rust_install():
+                http_archive(
+                    name = "rule_prefix__lazy_static__1_4_0",
+                    build_file = Label("//:BUILD.lazy_static-1.4.0.bazel"),
+                    sha256 = "e2abad23fbc42b3700f2f279844dc832adb2b2eb069b2df918f455c4e18cc646",
+                    strip_prefix = "lazy_static-1.4.0",
+                    type = "tar.gz",
+                    url = "https://crates.io/api/v1/crates/lazy_static/1.4.0/download",
+                )
+
+            "# };
+
+        assert_eq!(output, expected_repository_rule);
+    }
+
+    #[test]
+    fn render_git_repository() {
+        let renderer = mock_renderer(true);
+
+        let mut output = Vec::new();
+
+        renderer
+            .render_workspaces(&mut output)
+            .expect("Error rendering");
+
+        let output = String::from_utf8(output).expect("Non-UTF8 output");
+
+        // TODO: Don't unconditionally load new_git_repository and http_archive
+        let expected_repository_rule = indoc! { r#"
+            # rules_rust crate_universe file format 1
+            # config hash 598
+            
+            load("@bazel_tools//tools/build_defs/repo:git.bzl", "new_git_repository")
+            load("@bazel_tools//tools/build_defs/repo:http.bzl", "http_archive")
+            
+            def pinned_rust_install():
+                new_git_repository(
+                    name = "rule_prefix__lazy_static__1_4_0",
+                    strip_prefix = "",
+                    build_file = Label("//:BUILD.lazy_static-1.4.0.bazel"),
+                    remote = "https://github.com/rust-lang-nursery/lazy-static.rs.git",
+                    commit = "421669662b35fcb455f2902daed2e20bbbba79b6",
+                )
+
+            "# };
+
+        assert_eq!(output, expected_repository_rule);
+    }
+
+    #[test]
+    fn render_defs_bzl() {
+        let crate_context = testing::lazy_static_crate_context(false);
+
+        let normal_deps = btreemap! {
+            crate_context.pkg_name.clone() => crate_context.pkg_version.clone(),
+        };
+
+        let renderer = Renderer::new(
+            RenderConfig {
+                repo_rule_name: String::from("rule_prefix"),
+                crate_registry_template: String::from(
+                    "https://crates.io/api/v1/crates/{crate}/{version}/download",
+                ),
+                rules_rust_workspace_name: String::from("rules_rust"),
+            },
+            String::from("598"),
+            vec![crate_context],
+            Dependencies {
+                normal: normal_deps,
+                build: BTreeMap::new(),
+                dev: BTreeMap::new(),
+            },
+            btreemap! {
+                String::from("//some:Cargo.toml") => btreeset!{ String::from("lazy_static") },
+            },
+        );
+
+        let mut output = Vec::new();
+
+        renderer
+            .render_helper_functions(&mut output)
+            .expect("Error rendering");
+
+        let output = String::from_utf8(output).expect("Non-UTF8 output");
+
+        // TODO: Have a single function with kwargs to enable each kind of dep, rather than multiple functions.
+        let expected_repository_rule = indoc! { r#"
+        CRATE_TARGET_NAMES = {
+            "lazy_static": "@rule_prefix__lazy_static__1_4_0//:lazy_static",
+        }
+
+        def crate(crate_name):
+            """Return the name of the target for the given crate.
+            """
+            target_name = CRATE_TARGET_NAMES.get(crate_name)
+            if target_name == None:
+                fail("Unknown crate name: {}".format(crate_name))
+            return target_name
+
+        def all_deps():
+            """Return all standard dependencies explicitly listed in the Cargo.toml or packages list."""
+            return [
+                crate(crate_name) for crate_name in [
+                    "lazy_static",
+                ]
+            ]
+
+        def all_proc_macro_deps():
+            """Return all proc-macro dependencies explicitly listed in the Cargo.toml or packages list."""
+            return [
+                crate(crate_name) for crate_name in [
+                ]
+            ]
+
+        def crates_from(label):
+            mapping = {
+                "//some:Cargo.toml": [
+                    crate("lazy_static"),
+                ],
+            }
+            return mapping[_absolutify(label)]
+
+        def dev_crates_from(label):
+            mapping = {
+                "//some:Cargo.toml": [
+                ],
+            }
+            return mapping[_absolutify(label)]
+
+        def build_crates_from(label):
+            mapping = {
+                "//some:Cargo.toml": [
+                ],
+            }
+            return mapping[_absolutify(label)]
+
+        def proc_macro_crates_from(label):
+            mapping = {
+                "//some:Cargo.toml": [
+                ],
+            }
+            return mapping[_absolutify(label)]
+
+        def _absolutify(label):
+            if label.startswith("//") or label.startswith("@"):
+                return label
+            if label.startswith(":"):
+                return "//" + native.package_name() + label
+            return "//" + native.package_name() + ":" + label
+        "# };
+
+        assert_eq!(output, expected_repository_rule);
+    }
+
+    #[test]
+    fn render_crate_build_file() {
+        let renderer = mock_renderer(true);
+
+        let repository_dir = tempfile::TempDir::new().unwrap();
+        renderer.render_crates(repository_dir.as_ref()).unwrap();
+
+        let build_file_path = repository_dir
+            .as_ref()
+            .join("BUILD.lazy_static-1.4.0.bazel");
+        let build_file_contents = std::fs::read_to_string(&build_file_path)
+            .expect(&format!("Failed to read {}", build_file_path.display()));
+
+        assert_eq!(
+            indoc! { r#"
+                # buildifier: disable=load
+                load(
+                    "@rules_rust//rust:defs.bzl",
+                    "rust_binary",
+                    "rust_library",
+                    "rust_proc_macro",
+                    "rust_test",
+                )
+
+                # buildifier: disable=load
+                load("@bazel_skylib//lib:selects.bzl", "selects")
+
+                package(default_visibility = [
+                    "//visibility:public",
+                ])
+
+                licenses([
+                    "restricted",  # no license
+                ])
+
+                # Generated targets
+
+                # buildifier: leave-alone
+                rust_library(
+                    name = "lazy_static",
+                    deps = [
+                    ],
+                    srcs = glob(["**/*.rs"]),
+                    crate_root = "src/lib.rs",
+                    edition = "2015",
+                    rustc_flags = [
+                        "--cap-lints=allow",
+                    ],
+                    data = glob(["**"], exclude=[
+                        # These can be manually added with overrides if needed.
+
+                        # If you run `cargo build` in this dir, the target dir can get very big very quick.
+                        "target/**",
+
+                        # These are not vendored from the crate - we exclude them to avoid busting caches
+                        # when we change how we generate BUILD files and such.
+                        "BUILD.bazel",
+                        "WORKSPACE.bazel",
+                        "WORKSPACE",
+                    ]),
+                    version = "1.4.0",
+                    tags = [
+                        "cargo-raze",
+                        "manual",
+                    ],
+                    crate_features = [
+                    ],
+                    aliases = select({
+                        # Default
+                        "//conditions:default": {
+                        },
+                    }),
+                )
+
+            "# },
+            build_file_contents,
+        );
+    }
 }
